+++ backend/src/main/java/moaon/backend/project/dto/ProjectQueryCondition.java
@@ -1,4 +1,4 @@
-package moaon.backend.project.repository;
+package moaon.backend.project.dto;
 
 public record ProjectQueryCondition(
         String search
