+++ backend/src/main/java/moaon/backend/project/repository/ProjectRepositoryCustom.java
@@ -2,6 +2,7 @@ package moaon.backend.project.repository;
 
 import java.util.List;
 import moaon.backend.project.domain.Project;
+import moaon.backend.project.dto.ProjectQueryCondition;
 
 public interface ProjectRepositoryCustom {
