+++ backend/src/main/java/moaon/backend/project/repository/ProjectRepositoryImpl.java
@@ -7,6 +7,7 @@ import com.querydsl.jpa.impl.JPAQueryFactory;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import moaon.backend.project.domain.Project;
+import moaon.backend.project.dto.ProjectQueryCondition;
 import org.springframework.stereotype.Repository;
 import org.springframework.util.StringUtils;
