+++ backend/src/main/java/moaon/backend/project/service/ProjectService.java
@@ -7,8 +7,8 @@ import moaon.backend.global.exception.custom.ErrorCode;
 import moaon.backend.love.repository.LoveRepository;
 import moaon.backend.project.domain.Project;
 import moaon.backend.project.dto.ProjectDetailResponse;
+import moaon.backend.project.dto.ProjectQueryCondition;
 import moaon.backend.project.dto.ProjectSummaryResponse;
-import moaon.backend.project.repository.ProjectQueryCondition;
 import moaon.backend.project.repository.ProjectRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
