+++ backend/src/test/java/moaon/backend/project/repository/ProjectRepositoryImplTest.java
@@ -7,6 +7,7 @@ import java.util.List;
 import moaon.backend.fixture.RepositoryTestHelper;
 import moaon.backend.global.config.QueryDslConfig;
 import moaon.backend.project.domain.Project;
+import moaon.backend.project.dto.ProjectQueryCondition;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
